Move monomorphisation code to new crate
