+++ src/quill_link/Cargo.toml
@@ -5,5 +5,4 @@ authors = ["thirdsgames <thirdsgames2018@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-quill_llvm = { path = "../quill_llvm" }
 quill_target = { path = "../quill_target" }
