+++ src/quill_llvm/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 
 [dependencies]
 inkwell = { git = "https://github.com/TheDan64/inkwell", branch = "master", features = ["llvm11-0"] }
-guess_host_triple = "^0.1"
 quill_common = { path = "../quill_common" }
 quill_parser = { path = "../quill_parser" }
 quill_index = { path = "../quill_index" }
@@ -14,6 +13,7 @@ quill_type = { path = "../quill_type" }
 quill_type_deduce = { path = "../quill_type_deduce" }
 quill_mir = { path = "../quill_mir" }
 quill_target = { path = "../quill_target" }
+quill_monomorphise = { path = "../quill_monomorphise" }
 dunce = "1.0.1"
 
 [dev-dependencies]
